+++ include/backend/cuda/ops/initializer.h
@@ -10,22 +10,22 @@ template <Device D> struct RandomImpl;
 
 template <>
 struct ZerosImpl<Device::CUDA> {
-    static Tensor execute(const std::vector<int>& shape, DataType dtype);
+    static Tensor execute(const std::vector<int64_t>& shape, DataType dtype);
 };
 
 template <>
 struct OnesImpl<Device::CUDA> {
-    static Tensor execute(const std::vector<int>& shape, DataType dtype);
+    static Tensor execute(const std::vector<int64_t>& shape, DataType dtype);
 };
 
 template <>
 struct FillImpl<Device::CUDA> {
-    static Tensor execute(const std::vector<int>& shape, DataType dtype, float value);
+    static Tensor execute(const std::vector<int64_t>& shape, DataType dtype, float value);
 };
 
 template <>
 struct RandomImpl<Device::CUDA> { 
-    static Tensor execute(const std::vector<int>& shape, DataType dtype,float min,float max);
+    static Tensor execute(const std::vector<int64_t>& shape, DataType dtype,float min,float max);
 };
 
 extern template struct ZerosImpl<Device::CUDA>;
