+++ include/backend/sycl/ops/transpose.h
@@ -13,7 +13,7 @@ struct TransposeImpl;
 template <>
 struct TransposeImpl<Device::SYCL>{
     static void execute(Tensor& a);
-    static Tensor execute(Tensor& a, std::initializer_list<int> axes);
+    static Tensor execute(Tensor& a, std::initializer_list<int64_t> axes);
 };
 
 extern template struct TransposeImpl<Device::SYCL>;
