+++ include/core/tensor.h
@@ -15,8 +15,8 @@ public:
     virtual size_t numel() const = 0;
     virtual void copy_to(void* dst) const = 0;                      // 同设备
     virtual const void* data() const = 0;
-    virtual void reshape(std::vector<int>& newshape)=0;
-    virtual void reshape(std::initializer_list<int> newshape)=0;
+    virtual void reshape(std::vector<int64_t>& newshape)=0;
+    virtual void reshape(std::initializer_list<int64_t> newshape)=0;
     virtual std::unique_ptr<TensorImpl> clone() const = 0;
     virtual std::shared_ptr<ContextImpl> context() const = 0;
 };
@@ -26,7 +26,7 @@ private:
     size_t m_numel;
     Device m_device;
     DataType m_dtype;
-    std::vector<int> m_shape;
+    std::vector<int64_t> m_shape;
     std::shared_ptr<TensorImpl> m_impl;
 public:
     void* data();
@@ -46,20 +46,20 @@ public:
     Tensor();
     Tensor(const Tensor& other);            // 拷贝构造函数
     Tensor(Tensor&& other) noexcept;        // 移动构造函数
-    Tensor(std::vector<int> shape,DataType dtype);
-    Tensor(std::initializer_list<int> shape,DataType dtype);
-    Tensor(std::vector<int> shape,DataType dtype,Device device);
-    Tensor(std::initializer_list<int> shape,DataType dtype,Device device);
+    Tensor(std::vector<int64_t> shape,DataType dtype);
+    Tensor(std::initializer_list<int64_t> shape,DataType dtype);
+    Tensor(std::vector<int64_t> shape,DataType dtype,Device device);
+    Tensor(std::initializer_list<int64_t> shape,DataType dtype,Device device);
     // 从外部指针创建张量
-    Tensor(void* ptr, std::vector<int> shape,DataType dtype,Device device);
-    Tensor(void* ptr,std::initializer_list<int> shape,DataType dtype,Device device);// 深拷贝ptr的内容
+    Tensor(void* ptr, std::vector<int64_t> shape,DataType dtype,Device device);
+    Tensor(void* ptr,std::initializer_list<int64_t> shape,DataType dtype,Device device);// 深拷贝ptr的内容
 
     // 从stl容器创建张量,vector
     template<typename T>
-    Tensor(std::vector<T>& vec,std::initializer_list<int> shape);
+    Tensor(std::vector<T>& vec,std::initializer_list<int64_t> shape);
 
     template<typename T>
-    Tensor(std::vector<T>& vec, std::vector<int> shape);
+    Tensor(std::vector<T>& vec, std::vector<int64_t> shape);
 
     Tensor& operator=(const Tensor& other); // 拷贝赋值运算符
     Tensor& operator=(Tensor&& other) noexcept;  // 移动赋值运算符
@@ -93,12 +93,12 @@ public:
     
 
     // 支持负数下标的访问：int val = shape(-1);
-    int shape(int i);
-    int shape(int i)const;
-    std::vector<int> shape()const;
+    int64_t shape(int i);
+    int64_t shape(int i)const;
+    std::vector<int64_t> shape()const;
     // reshape,不改变数据排列
-    void reshape(std::vector<int>& newshape);
-    void reshape(std::initializer_list<int> newshape);
+    void reshape(std::vector<int64_t>& newshape);
+    void reshape(std::initializer_list<int64_t> newshape);
 
     std::shared_ptr<TensorImpl> get_impl() const;
 
@@ -148,19 +148,19 @@ public:
     
     // 元素访问
     template<typename T>
-    T at(std::initializer_list<int> idxs);
+    T at(std::initializer_list<int64_t> idxs);
 
     template<typename T>
-    T operator[](std::initializer_list<int> idxs);
+    T operator[](std::initializer_list<int64_t> idxs);
 
     // 常用静态函数
-    static Tensor Zeros(std::initializer_list<int> shape, DataType dtype = DataType::FLOAT32);
-    static Tensor Zeros(std::vector<int> shape, DataType dtype = DataType::FLOAT32);
-    static Tensor Ones(std::initializer_list<int> shape, DataType dtype = DataType::FLOAT32);
-    static Tensor Ones(std::vector<int> shape, DataType dtype = DataType::FLOAT32);
-    static Tensor Fill(std::initializer_list<int> shape, float value, DataType dtype = DataType::FLOAT32);
-    static Tensor Fill(std::vector<int> shape, float value, DataType dtype);
-    static Tensor Random(std::initializer_list<int> shape, float min = 0.0f, float max = 1.0f, DataType dtype = DataType::FLOAT32);
-    static Tensor Random(std::vector<int> shape, float min = 0.0f, float max = 1.0f, DataType dtype = DataType::FLOAT32);
+    static Tensor Zeros(std::initializer_list<int64_t> shape, DataType dtype = DataType::FLOAT32);
+    static Tensor Zeros(std::vector<int64_t> shape, DataType dtype = DataType::FLOAT32);
+    static Tensor Ones(std::initializer_list<int64_t> shape, DataType dtype = DataType::FLOAT32);
+    static Tensor Ones(std::vector<int64_t> shape, DataType dtype = DataType::FLOAT32);
+    static Tensor Fill(std::initializer_list<int64_t> shape, float value, DataType dtype = DataType::FLOAT32);
+    static Tensor Fill(std::vector<int64_t> shape, float value, DataType dtype);
+    static Tensor Random(std::initializer_list<int64_t> shape, float min = 0.0f, float max = 1.0f, DataType dtype = DataType::FLOAT32);
+    static Tensor Random(std::vector<int64_t> shape, float min = 0.0f, float max = 1.0f, DataType dtype = DataType::FLOAT32);
 };
